Fix test project build breaks. Entity lease tests need to be updated for new API

+++ Headers/Compositor/EntityArena.h
@@ -14,7 +14,7 @@ class EntityArena {
  public:
   EntityArena(uint8_t* base, size_t maxSize, bool reader);
 
-  PresentationEntity* emplaceEntity(const Entity& entity, size_t parentIndex);
+  PresentationEntity* emplaceEntity(const Entity& entity, int32_t parentIndex);
 
   const PresentationEntity& operator[](size_t index) const;
